% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf2vcf.R
\name{maf2vcf}
\alias{maf2vcf}
\title{Converts MAF to a VCF files.}
\usage{
maf2vcf(maf)
}
\arguments{
\item{infile}{Path MAF file}
}
\value{
A VCF file
}
\description{
This script will take the MAF file and convert it to a minimal
VCF file. This will REMOVE large amounts of VITAL information.
}
